requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.5
django-celery-beat==2.4.0
django-crispy-forms==1.14.0
django-timezone-field==5.0
djangorestframework==3.14.0
flower==1.2.0
humanize==4.4.0
idna==3.4
kombu==5.2.4
pip-system-certs==4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
pycodestyle==2.10.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.2
requests==2.28.2
secure-smtplib==0.1.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1